mainwindow,playlistwindow,helpers: move filtering to helpers
Filter items as they come into the playlist, rather than opened.
We have a variety of ways of adding files and folders, so better
to be safe and catch them in one place.  This filtering involves
expanding directories ourselves where possible, since we don't want
unplayable items.

Also includes image file types which were not in the previous commit.

As pointed out in #174.