manager: open with the optionally changed url
Actually fetch the url of the first item before playing it.  This can
sometimes change by ourselves.  Of course, if mpv wants to expand that
as well, there's no stopping that.

Currently needed because we used to rely on mpv expanding folders, now
we do it instead.

As mentioned in #178.