settingswindow: make hide panels fullscreen persistent
Or rather, remember it when it's changed through the playlist context
menu. i.e. emit the accepted settings data map after it's changed.

If I were to rewrite this, I'd use a shared pointer between the settings
window and flow instead of pumping out a bunch of data whenever.  This
is probably a pointless microoptimisation however.