.pro,mainwindow: provide more detailed version info
The build will try hard to determine if it is on a tag, extract its
version label, and not indicate that it is a development version where
possible.  The idea here is that you can be lazy and still get away with
it if you're in a git repo.  :)

Package maintainers who wish to do their own labelling for a particular
release can now pass MPCQT_VERSION=whatever in a straightforward way, or
the old way through a escape-coded define to qmake, and the build will
use it without much parsing.  Therefore if one were to download the next
release zip and your old package script runs unedited, the build will
pick it up you should not notice anything peculiar.

If it's a zipfile and you haven't passed any version in the new way
through MPCQT_VERSION=YY.MM or the old way through a MPCQT_VERSION_STR
define, then you'll get the usual generic version label.  No biggie.