devicemanager: add to project
Device manager is a suite of classes designed to track device insertion
and removals.

As of the commit, the WinApi half written (it does an initial scan but
does not update), and the rest are skeletons.  I have a udisks2 dbus
interface lying around that probably will become the backbone of the
unix interface.