devicemanager: implement udisks2 interface
Use information provided by the udisks2 interface to determine available
"discs".  This is optional, because it won't keep probing if the service
is not present.

This information is kept reasonably up to date across device addition/
removals etc, at least internally.  The front-end code however does not
at present track this information, so a device's entry will stop working
if it is unplugged at all, and new devices will not be shown.  This
deviation amounts to wiring stuff up and triggering the updateDiscList
function.

It also does not request a mount point, this needs to be exposed to the
general interface.  I can confirm that dvds work if some other program
mounts it (even after startup).

I made some changes to the general interface, so compiling on Windows
needs to be checked. (it probably works.)