mainwindow: add shortcut for escaping fullscreen
At present, escape is used for exiting out of the search field.  A
nearby but unused key is the backtick key next to the letter 1, so use
that instead.  Perhaps this way of escaping out of the search field
could be implemented differently to free up this shortcut, perhaps as an
event filter or so.

This is customizable in the action editor.  Note that the action editor
has a quirk whereby the shortcut editor has to be deselected for the
selected shortcut to be written out to the model, which is easy to
forget about.

Fixes #194.