devicemanager: win: listen for device add/removals
Use a device listener window that forwards rescan requests upon every
addition/removal message.  Receiving a lot of events in quick succession
could be potentially expensive (yanking a 100 port hub for example), so
we also make a small attempt to delay these requests until the
application is idle with a zero-length timer.