logger,logwindow: add logging framework to project
Create a logging framework with goal of showing debugging messages when
run outside a terminal.  Add the three log-related classes described
below and wire them into the application.  Showing the log window is
exposed under the main window's View menu.  The debugging box in the
settings window is moved to its own page, with associated options added.
Saved geometries will be reset because old geometry files have no
information about the log window and are therefore technically
incomplete.

Logger is a log message receiver designed to sit in its own thread,
buffer log messages, and keep a log file.  It is capable of descriptive
message labelling, storing messages, delayed flushes, and intercepting
debug messages from Qt.

LogStream is a replacement for qDebug() that forwards its contents to
Logger inside its destructor.  It is capable of serializing variants and
other daya with the usual c++ stream operator.  It is more expensive
than the log(s) functions in Logger.

LogWindow is a window to display in a a plain text document the log
messages received from Logger.  It is capable of setting its history to
a custom limit, its contents cleared, and copying its contents to the
clipboard and/or a file.

Fixes #196.