settingswindow: add audio autoload section
Spin out the audio device selection to its own page and add an autoload
audio section to it.  This is similar to the layout in mpc-be.
Untested, but it should work.

The old output page looks a bit bare, perhaps there's some overlooked
options that can be placed there.

Fixes #213.