manager: implement after playback->play next.
Split up the playNextFile function to allow file url injection.  Use the
now playling url to fetch the next file in the folder.

Skips on to the next track in the playlist at past the last file.
Note: this is a quick way to destroy your playlist.  Be careful!

Fixes #227.