http: add partial http server
Adds a mpc-hc compatible(?) server.  I expect it to differ in
compatibility because I lack the means to thoroughly test it.

Partially written.  What's missing is at least the browser, info, and
variables pages.  So the server will need to know about player state as
well.  And command.html needs a gigantic translation-layer to my actions
for all the implemented command-ids, so there'll be boilerplate one way
or another.  Subject to change.

The only reason this is getting pushed in this unfinished state is
because I'm going on holidays for two weeks so there's an outside chance
I may flesh it out on my laptop.  At least I'll have some videos with me
this time.