ipc/http: populate command list
Populate command list with implemented functions that we should be able
to do.  Use two containers, one which orders by insertion order and one
which orders by id.

We were going to create this list in mainwindow using a map of id-to-
QAction and pass it around, but that left a bad taste in my mouth
because it would be putting part of the mpc-hc http front-end data in my
gui module.  So hard-wire the text instead of simply extracting the ui
text, and (like a good modularized programmer) we're going to use a
bunch of signals for this instead of invoking a foreign QAction.  On
another day we'll wire it all up in main and patch in new slots where
needed.  Ta ta now.