http: emit signals on command
This covers stuff which we implement.  There's probably more that I
missed, but this should do for now.  Hopefully something like a mpc-hc
remote should work now, but don't quote me on that.

Note that command.html isn't complete yet because it's supposed to
handle seek events as well.  This is used by the inbuilt interface that
mpc-hc has, but we don't have.  So until that gets done I see no reason
to bother with it.