main: use xcb under linux et al
Use xcb to restore expected behaviour as experienced on other operating
systems.  Window management is a broken mess otherwise: it's easily
possible to open videos with half the window offscreen despite our
attempts to place ourselves in the right position, and other things such
as windows not opening where we tell them to.