make-release-win: make more future proof
Libicu dlls seems to have changed by the time I make a release, so copy
those into the zip with a glob.  This should make it more future proof.

However, it seems that there is the odd elusive dll that changes every
now and then (such as libpcre), to the point that running this script
inside a Windows build server would be troublesome.  We still need a
more automated way to do this.