main,settingswindow: add wayland override
Wayland is a nice experience with a silky-smooth 60fps, BUT no run-time
centering is a bit of a deal-breaker for this application.

Try it: select use of wayland in the settings dialog, restart, have a
single FHD monitor like the majority of people, have the window anywhere
on the desktop that isn't (0,0), leave the video zoom at the default
1:1, and play a 1080p video.  Every-frame-is-perfect is nice when you
can actually see it.

That said, some people don't care about that.

Closes #72.