mainwindow: rename mpvHost() -> dockHost()
Rename the mpvHost() function because it is now used as a different
purpose than its old name.  I experiemented with changing docks from the
nested dock to the main window under wayland; this revealed that the
function is only ever used to obtain access to the nested dock
container, not so much with the fact that mpv is nested inside.