mainwindow: add media keys on windows
Windows needs key interception to process media key events, while Linux
uses MPRIS.  Presumably this code is needed on Mac too, but I don't have
a machine to test that.

Play and Pause button aren't on this keyboard that I passed through to
the VM, so those are untested.  PlayPause, Stop, Previous and Next are
though.

Fixes #34.