mainwindow: hide certain items in the view menu
From my testing in mpc-hc, these do nothing or are never enabled with
regular files.

Subresync: Not needed.  Get better rips if this is a problem for you.

Capture: There are specific applications that are better tailored for
this use case, such as xine and kamera.  And besides, there are several
problems with developing this feature.  (1) I never had any experience
using capture cards with mpc-hc when I had a TV card 15 years ago; I
used an app bundled with the TV card, which leads me to believe that
applications specifically designed for television watching provide a
better experience.  (2) I doubt I want to spend the investment in making
TV cards work with libmpv+w_scan anyway.  (3) I don't have a TV point in
this room.  (4) If this feature is supposed to be for webcams, then too
bad because mpc-hc also doesn't pick up my webcam.

Navigation: Doesn't seem to turn on even with chaptered files.  The
existing ui is sufficient for navigating these.  And if this a feature
that is enabled with dvds/bds, you really should be using xine for
those; but I can't test if this is enabled with them because all my
discs are packed in a box and the last shop that sold discs in my town
(Sanity) closed up shop several months ago.  Additionally, you should be
ripping your discs.

Fixes #125.