mpvwidget: patch around wayland cursor hide bug
Check that main window is the active window before attempting to hide
the cursor.  Wayland doesn't seem to understand the mouse is over a
different widget if it happens to be in a different window.

Unfortunately this means that mouse cursor will not always hide after a
delay when it gets placed over the video area while another obscuring
window is active, but you probably want to always see the cursor when
you have a different window active anyway.  At least this fix doesn't
feel all that odd.

But that's only if you have the player active.   Suppose you are active
in a different application entirely, with the player in a corner of the
screen; hiding the mouse cursor over the video area will no longer work
in that case.  Too bad!  Use X11 if you want all of the app to work!