main,mainwindow,manager: set play after default only once
Use a set-once flag to control the propagation of the play after setting
in the settings window to the rest of the program.  Without this fix, it
would receive the signal to change the playafter every time option soon
after playback, and requests to play next folder after every file would
revert to whatever is set in the settings window.

Fixes #129.