Play next/previous video in folder
By default, PgDown and PgUp will play the next/previous video in folder
if there's no more chapter and there's only one file in the playlist.

With the renamed "Play next file in folder when there is only one item
in the playlist" option enabled, the next file in folder will play
without pressing anything when the end of the file is reached and the
playlist has only the current file.