Recents and eof fixes for play next/previous in folder
Don't close the file at the end, just pause the player, like MPC-HC.
This allows manually opening the next file when the playing stopped at
the end, seeking back, and being able to see we've reached the end.

Wait for playlength to change before updating recents.
This avoids using the length of the previous file for the current one or
using -1 as the length.
Since the length is set to -1 when opening a file before being set to
its actual size, the recents are updated even if the next file has the
same length (first to -1 then to real length).