manager: fix early finish
Fake a file end event when a file does not properly initialize.

Consider the case where a stream is opened but fails to properly
initialize such as the url https://chiru.no:8000/stream.mp3.  In these
scenarios libmpv would finish playback before it starts, and the player
would be stuck in the Buffering state.  We need to progress to the next
file in the playlist.  We need to get out of the Buffering state when
there is no file after us in the playlist.  So send a stopped state
event and jump to the code that handles an end of file event.  This
mimics the behviour of mpc-hc from my tests.