Don't convert eof-reached to bool
eof-reached is the first watched property notification that a new file
is being opened.
On file change, eof-reached, which is then an empty string, can't be
converted to bool, so it's set to default (false). Thus we can't know if
it's a new file or a file being played again after reaching the end.