Recreate the playlist when playing the next file in folder
This prevents recents from having duplicate uuids (as
Playlist::replaceItem reuses the uuid for the first file), which breaks
them.
Ideally, this would be done in Playlist::replaceItem to also handle
removing an archive item when it doesn't contain any playable file. But
there's also the fact that PlaybackManager::mpvw_playlistChanged isn't
called when an archive contains only one file.