Close playlistWindow with MainWindow
Otherwise, it makes the app hang when closing with QApplication::quit().
That doesn't mean we want to use QApplication::quit() as it there are still cases where it makes the app hang on close.

We need to restore the state of the playlist window (shown/hidden) action after closing it so that it gets remembered for the next start.