Flush pending messages to disk on fatal message
Also flush any pending messages to the log file.  Most users on Windows
would not see the pending message dump on stderr, and depend on the file
to retrieve the log.  This may make fatal errors reported by qFatal more
complete.  However, if the logs are not set to immediate mode and a
segfault happens between dumps to disk, messages will still get missed.

To do this, we reword the qt message handler to file the fatal log
message normally, so that it will get picked up by the final message
dump, for both file and console.  And change the fatalMessage procedure
call into a blocking call on the log thread, so that we aren't trying to
dump pendingMessages while the log thread is writing to it... hopefully
the application is still in a good enough state to do blocking calls.
Because the fatalMessage handler is now in the log thread where it
should be, it can empty the pending message queue so that these messages
won't be written twice before the application quits.