Let selected tracks get reset when opened from recent files
If the current file has subtitle id 2 manually selected and we open a
recent file which has no manually selected subtitle (id = -1), it'll end
up having subtitle 0 (or 2 if it has at least two) selected, thus
bypassing auto track selection.
This applies to audio and video tracks too.

So we let the tracks be set to -1 and add a safety check in MainWindow.

Setting the tracks in Flow:mainwindow_recentOpened for recent files when
rememberFilePosition is true is probably not needed as they are already
set in Flow::manager_startingPlayingFile.
But let's play it safe as this is a bug fix.