Save position of current file when playing new one from playlist
When we start playing a file from a playlist,
PlaybackManager::startPlayWithUuid gets directly triggered, so we need
to save the position when that happens.

We need to check if the url isn't empty because it gets reset when we
open the same file again from the playlist, so it would add an empty
entry in recent files.

We could probably stop calling Flow::updateRecentPosition in
Flow::manager_playLengthChanged if we call updateRecentPosition again at
the end of Flow::manager_startingPlayingFile, but it'll require more
testing to avoid regressions.