Rename .desktop file to use recommended appstream format
The id can only contain an hyphen in the last part, the other one has to
be converted to an underscore. And since Flatpak renames the .desktop
file to the id, we have to follow that format if we want to avoid
#337.
Follow-up to fb1a71af709f82f308797d9d9e3a3861a6bd79d2.