Use add-build-extensions in Flatpak manifest and replace FLATPAK_DEST
Using add-build-extensions instead of add-extensions ensures that
ffmpeg-full is available during the build too.
Along with CMake for the build, it somehow fixes H.264 and H.265
(hardware) decoding.
I noticed with the command `flatpak enter io.github.mpc_qt.mpc-qt bash`
(to be used while the Flatpak version is running) that /app/lib/ffmpeg
was empty, which means that ffmpeg-full wasn't available.