Remove audio volume from Options and move balance to audio page
Volume setting isn't very useful in the Options as it is usually set in
the main window.
And since 0e84a7fef2a2eb7bc40618e63a2b3816535503bd, when we open the
Playback settings, the volume has the same value as during the app
start.
This means that when we apply or close the Options with Ok, the main
window volume slider gets set to the settings volume.

Fixes #291.