Add a music mode
The purpose is to use all the available space for the playlist and make
it possible to keep a small window size.

This enables the playlist, makes it use the full width and height, and
prevents resizing, moving or undocking it.
As the playlist uses all the space (apart from the menus and buttons),
resizing the window resizes the playlist.
This also disables cover art to prevent the window from getting resized
if auto zoom is enabled.
A few buttons are also hidden: step backward and forward, and subtitles.

QApplication::processEvents() is needed to prevent the playlist from
playing catch-up with the window on resize.

Fixes #220.