Add a video preview to the seekbar
The video preview, enabled by default, is centered on the mouse cursor
and is restricted from extending beyond the window boundaries.
The background and text colors of the video preview can be customized
similarly to the regular tooltip, though changes only take effect after
a restart.
For audio files, a standard tooltip is used as reusing the same widget
for audio files would add complexity without adding value.
On Windows, it doesn't seem to show in fullscreen, like tooltips.

Fixes #4.