mainwindow: change shortcuts for volume up+down
Change shortcuts of volume up and volume down to ctrl+shift+up and
ctrl+shift+down.  Ever since #330 the disable aspect ratio shortcut and
the volume down shortcut have been the same.  The results was that the
key did nothing.  Mpc-hc has these shortcuts as Up and Down.  However,
using these keys as shortcuts breaks keyboard navigation of the
playlist.  So update these modifiers to the ones in mpc-hc, with extra
modifiers so as to not conflict with other entries.

The volume shortcut keycombo is now marked as not-translatable and so
its entry is gone from the translations.

Other things in mainwindow.ui file changed when I loaded it in.  Most
notably qtcreator's form editor keeps crying out for a resource file.
(Resources are now in the cmake file.)  I manually edited it to look at
the res file under the build directory.  Hope it doesn't break anything.
I tested it with a cmake build as described in the README and it all
still seems to work, so fingers crossed.