Always use pointing hand cursor for seekbar
If the mouse pointer was on the seekbar before loading the first file,
it wouldn't change to a pointing hand cursor even after moving it on the
seekbar.
This has to be in MainWindow instead of in DrawnSlider to work.
Follow-up to ac98edd434c56f1f6c35cb07788046ecf4f0fc30.