Add OpenSubtitles web search support
Implements the "Search..." entry in "File > Subtitle Database" to open a
browser web page with the search results for the current video on
opensubtitles.org.

The new OpenSubtitles API (which is required to download subtitles from
the player) limits usage to 5 downloads per day (or 10 when logged in).
This limit can be reached very quickly when we have to download multiple
subtitles to find a good one.
That significantly reduces the usefulness to add support for the API.

#110.