Implement language override to English in Player options
Makes it easier for translators to check the original strings while
correcting them.
Also allows users of partial translations to revert to English.

We may want to list supported languages and store the language code in a
playerLanguage setting (as the QComboBox indexes will change).
This would increase the burden of adding translations to the list.