Draw VideoPreview on the main window
Avoids problems with Wayland (VideoPreview non moveable) and on Windows
(VideoPreview not visible in fullscreen). Also fixes some issues and
quirks we had to work around.

We move the VideoPreview widgets outside the screen to hide them. That's
because after being made invisible, they can't be made visible again.

The code is hopefully more readable now.