Output stacktrace in log and handle SIGABRT
Uses addr2line, so it will only output the address on Windows.

On Linux, it will output the function names if it's not a debug build.

On Windows, drmemory's symquery can output function names, but Boost
doesn't (yet?) support it.

Killing the app will also output the stacktrace in the log, which is
useful in case it hangs.