nsis: Create file associations on Windows install
Uses @he3als' mpc-qt-assoc (install and uninstall scripts and file icon)
itself based on @rossy's mpv-install.
It's a little slow (20-30 seconds).
The user still has to change the default app for playing videos in the
Windows settings (the script does open the settings when in attended
mode, but we use it in unattended mode).
I've tweaked the script a little bit to capitalize MPC-QT's name and to
add both the default install directory and make it work when installed
elsewhere with the script in a subdirectory).