nsis: Skip file associations creation whenever possible
Adds a registry key to check if file associations have already been done
previously to speed-up installation.
We might want to show a custom message on the welcome page to let users
know that we detect and update previous versions, so that users don't
uninstall before installing a new version. Because uninstalling removes
file associations.