cmakelists: Add ccache support to speed-up incremental builds
Disabled by default, can be enabled with the CCACHE_SUPPORT option in
the IDE in the CMake Cache settings or in the project build kit CMake
settings.

Ccache uses hashes to check which files need to be rebuild.
It makes Ninja as fast as make despite rebuilding more files.
It also reduces the time needed to rebuild single files:
settingswindow.cpp takes half the time with both make and Ninja.

Original cmakelists code by Konstantin Kharlamov.