Remember streams position and show them in Recent files
For streams, the title is shown in Recent files instead of the file name.
Like for local files, loading a Favorite still overrides the position
from recents. This may not be what the user wants.

Follow-up to f8c7aab2d405d6e15538aff96d582a783a02fefa which checked for
local file url in updateRecentPosition when it should have checked for
empty url.