mainwindow: Use Escape key shortcut for "Finish Searching" only while searchbar opened
The Escape key shortcut is used by default for closing the playlist
searchbar.
But this shortcut only makes sense while the searchbar is opened.
So it's switched to "Finish Searching" while the searchbar is opened and
returned to the default or user set action once the searchbar is closed.

Follow-up to 0204f25bef98d65f194d224fecf1c9e29425d429.