Use a custom widget for the seekbar tooltip
Drawn on mainwindow like the video preview.

Avoids a recent (Qt?) bug where the mouse cursor doesn't always stay as
a pointing hand and then preventing from clicking on the seek bar.

Also makes it possible to show the tooltip in fullscreen on Windows.

However, in the "below seekbar" mode, the tooltip isn't visible if there
isn't enough vertical space available below the seekbar.
For instance, in minimal or compact mode.