actioneditor: Use more descriptive names for commands in Keys editor
As was mentioned in #21, the volume "Up" and "Down" commands are not
easy to find in the Keys editor.
A few other commands names could be more descriptive too.

So this replaces their names (what is shown to the user) - or the
"actions descriptive text" per the Qt documentation - by a more verbose
one but only in the Keys editor.