settingswindow: Regroup and fix subtitles shadow settings
Regroup subtitles shadow color and shadow offset and use a checkbox to
enable them like backcolor did.

The shadow offset is set to 1 by default so that the shadow is directly
visible when enabled.

"sub-shadow-color" is an alias for "sub-back-color".
So the backcolor setting was overriding the shadow color setting and
hiding the shadow if not enabled.

"Subtitles shadow" is more user-friendly and is what MPC-HC uses.

Fixes #582 (Bug: Subtitle shadows missing after program restart).