videopreview: Request minimum file size for yt-dlp
This is recommended in the yt-dlp documentation to get the absolute
minimum video file size.
The "worst" preset isn't specific enough and doesn't work for all video
sources. "worstvideo" is better but still isn't specific enough as we're
interested about the filesize.
This preset notably works better for bilibili (see #466, though that
issue isn't about the video preview).