Don't percent-encode filename for screenshot name and use stream title for streams
QUrl::toDisplayString() returns a percent-encoded string ("%5B" and
"%5D" instead of "[" and "]"), even when using the QUrl::FullyDecoded
format.
QUrl::fileName() doesn't have that problem and already uses that format
by default. It also returns the filename so we don't have to manually
remove the directory path.

Like for recent files, streams usually don't have a useful filename, so
their title is used instead.
If the title includes a slash, we have to replace it for *nix systems.

Fixes #600 (Keep the original file name when saving screenshot (don't
convert characters to percent-encoding character)).