Add option to append opened files to Quick Playlist
If a file is already playing, the new files are appended and played on
their turn. If no file is currently playing, the new files are appended
and played immediately. If the player is paused, the new files are just
appended.

Enabling this option disables the "Remember Quick Playlist" option by
default to prevent the Quick Playlist from getting bigger and bigger
without the user noticing. It's still possible to manually re-enable
that setting.

Fixes #393 (Append opened files to playlist, instead of replacing
current).