settingswindow: Increase maximum values for normal and large playback steps
Increases them from 60 seconds to 2 000 000 000 seconds (a little more
than 23 days), close to actual 64-bit integer limit.

Note that mpv actually uses a double for the maximum seek size, so there
would be no practical limit if we switched to that. Not that it would
make much sense.

Fixes #627 (Long skip max duration is 60 seconds).