mainwindow: Make Play / Pause action trigger Play button
Dynamically rename the Pause action to Play or Pause.

Make it play the selected playlist (or by default the last played) item
from the stopped state, like MPC-HC.

Since the Play button was focused on start, it was possible to trigger
it with the space bar, which is also the default Play / Pause action
key. But that stopped being the case if the user switched to fullscreen
first or clicked the Mute button.

Most or all keyboards combine the Play and Pause keys. The Pause media
key already triggered both Play and Pause and will now also trigger Play
from the stopped state, as will the combined Play / Pause media key.